FinalProject/requirements.txt
ollama
pytesseract
Pillow
pymupdf